From a set of email the task is to print a list containing only valid email addresses in lexicographical order.
Valid email addresses must follow these rules:

It must have the username@websitename.extension format type.
The username can only contain letters, digits, dashes and underscores (a-z)(A-Z)(0-9).
The website name can only have letters and digits (a-z)(A-Z)(0-9).
The extension can only contain letters (a-z)(A-Z).
The maximum length of the extension is 3.

Concept

A filter takes a function returning True or False and applies it to a sequence, returning a list of only those members of the sequence where the function returned True. A Lambda function can be used with filters.